improve duplicate code
